+++ Runtime/Memory/GarbageCollector.cpp
@@ -30,7 +30,7 @@ void Astral::GarbageCollector::Cleanup()
 #endif
 #endif
 
-	for (int i = memory.size() - 1; i >= 0; i--)
+	for (int i = (int)memory.size() - 1; i >= 0; i--)
 	{
 		//If a block has no references, it can be deleted
 		if (memory[i]->CanDelete())
@@ -55,10 +55,10 @@ void Astral::GarbageCollector::Cleanup()
 #endif
 
 	//Estimate how many kbs are in use. Since different types have a differnt size, this is only a minimum, but it will be good enough
-	unsigned int size = (danglingMemory.size() * sizeof(Type::atype_t)) / 1024;
+	unsigned int size = (unsigned int)(danglingMemory.size() * sizeof(Type::atype_t)) / 1024;
 	if (size >= CullLimit) //If we have too much memory, all of the data we can
 	{
-		for (int i = danglingMemory.size() - 1; i >= 0; i--)
+		for (int i = (int)danglingMemory.size() - 1; i >= 0; i--)
 		{
 			if (!danglingMemory[i]->isOnStack) //If this is not on the stack, then we know it is not in use and ready to be cleaned up
 			{
@@ -89,7 +89,7 @@ void Astral::GarbageCollector::RegisterDanglingPointer(Type::atype_t* data)
 
 void Astral::GarbageCollector::RemoveDanglingPointer(Type::atype_t* data)
 {
-	for (int i = danglingMemory.size() - 1; i >= 0; i--)
+	for (int i = (int)danglingMemory.size() - 1; i >= 0; i--)
 	{
 		if (danglingMemory[i] == data)
 		{
